+++ Products/CMFPlone/utils.py
@@ -1,4 +1,5 @@
 import re
+import string
 from types import ClassType
 from os.path import join, abspath, split
 
@@ -16,12 +17,17 @@ from zope.i18n import translate
 from zope.publisher.interfaces.browser import IBrowserRequest
 
 import OFS
-from AccessControl import getSecurityManager, Unauthorized
+from AccessControl import getSecurityManager
 from AccessControl import ModuleSecurityInfo
+from AccessControl import Unauthorized
+from AccessControl.ZopeGuards import guarded_getattr
+from Acquisition import aq_base
 from Acquisition import aq_get
-from Acquisition import aq_base, aq_inner, aq_parent
+from Acquisition import aq_inner
+from Acquisition import aq_parent
 from App.Common import package_home
 from App.ImageFile import ImageFile
+from collections import Mapping
 from DateTime import DateTime
 from DateTime.interfaces import DateTimeError
 from Products.CMFCore.permissions import ManageUsers
@@ -670,3 +676,74 @@ def set_own_login_name(member, loginname):
     if not secman.checkPermission(ManageUsers, member):
         raise Unauthorized('You can only change your OWN login name.')
     pas.updateLoginName(member.getId(), loginname)
+
+
+class _MagicFormatMapping(Mapping):
+    """
+    Pulled from Jinja2
+
+    This class implements a dummy wrapper to fix a bug in the Python
+    standard library for string formatting.
+
+    See http://bugs.python.org/issue13598 for information about why
+    this is necessary.
+    """
+
+    def __init__(self, args, kwargs):
+        self._args = args
+        self._kwargs = kwargs
+        self._last_index = 0
+
+    def __getitem__(self, key):
+        if key == '':
+            idx = self._last_index
+            self._last_index += 1
+            try:
+                return self._args[idx]
+            except LookupError:
+                pass
+            key = str(idx)
+        return self._kwargs[key]
+
+    def __iter__(self):
+        return iter(self._kwargs)
+
+    def __len__(self):
+        return len(self._kwargs)
+
+
+class SafeFormatter(string.Formatter):
+
+    def __init__(self, value):
+        self.value = value
+        super(SafeFormatter, self).__init__()
+
+    def get_field(self, field_name, args, kwargs):
+        """
+        Here we're overridding so we can use guarded_getattr instead of
+        regular getattr
+        """
+        first, rest = field_name._formatter_field_name_split()
+
+        obj = self.get_value(first, args, kwargs)
+
+        # loop through the rest of the field_name, doing
+        #  getattr or getitem as needed
+        for is_attr, i in rest:
+            if is_attr:
+                obj = guarded_getattr(obj, i)
+            else:
+                obj = obj[i]
+
+        return obj, first
+
+    def safe_format(self, *args, **kwargs):
+        kwargs = _MagicFormatMapping(args, kwargs)
+        return self.vformat(self.value, args, kwargs)
+
+
+def safe_format(inst, method):
+    """
+    Use our SafeFormatter that uses guarded_getattr for attribute access
+    """
+    return SafeFormatter(inst).safe_format